Flask-Backend-Markdown/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-docx==1.0.1
html2text==2024.2.26
Werkzeug==2.3.7
gunicorn==21.2.0
pywin32==306; sys_platform == 'win32'